requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.1.0
execnet==1.9.0
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
playwright==1.29
pluggy==1.0.0
py==1.11.0
pyee==9.0.4
pytest==7.2.0
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-playwright==0.3.0
pytest-xdist==3.1.0
python-slugify==7.0.0
requests==2.28.1
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
pytest-asyncio
pylint-fail-under
